Cleanup `compositor_render`
