+++ compositor_pipeline/src/pipeline.rs
@@ -7,13 +7,14 @@ use std::time::Duration;
 use compositor_render::error::{
     ErrorStack, InitRendererEngineError, RegisterRendererError, UnregisterRendererError,
 };
-use compositor_render::renderer::RendererOptions;
 use compositor_render::scene::Component;
+use compositor_render::web_renderer::WebRendererInitOptions;
+use compositor_render::RegistryType;
+use compositor_render::RendererOptions;
 use compositor_render::{error::UpdateSceneError, Renderer};
 use compositor_render::{
     scene, EventLoop, Framerate, InputId, OutputId, RendererId, RendererSpec, Resolution,
 };
-use compositor_render::{RegistryType, WebRendererOptions};
 use crossbeam_channel::unbounded;
 use ffmpeg_next::Packet;
 use log::{error, warn};
@@ -72,7 +73,7 @@ pub struct Pipeline<Input: PipelineInput, Output: PipelineOutput> {
 pub struct Options {
     pub framerate: Framerate,
     pub stream_fallback_timeout: Duration,
-    pub web_renderer: WebRendererOptions,
+    pub web_renderer: WebRendererInitOptions,
 }
 
 impl<Input: PipelineInput, Output: PipelineOutput> Pipeline<Input, Output> {
