+++ compositor_render/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod error;
-pub mod renderer;
 pub mod scene;
 
 pub(crate) mod registry;
@@ -8,21 +7,28 @@ pub(crate) mod utils;
 pub(crate) mod wgpu;
 
 mod event_loop;
-mod sync_renderer;
+mod state;
 mod types;
 
 pub use event_loop::EventLoop;
 pub use types::*;
 
-pub use transformations::web_renderer::{
-    WebRendererOptions, EMBED_SOURCE_FRAMES_MESSAGE, GET_FRAME_POSITIONS_MESSAGE,
-    UNEMBED_SOURCE_FRAMES_MESSAGE,
-};
-
-pub type Renderer = sync_renderer::SyncRenderer;
 pub use registry::RegistryType;
-pub use sync_renderer::RendererSpec;
+pub use state::Renderer;
+pub use state::RendererOptions;
+pub use state::RendererSpec;
 
-pub use transformations::image_renderer::{ImageSpec, ImageSrc, ImageType};
-pub use transformations::shader::ShaderSpec;
-pub use transformations::web_renderer::{WebEmbeddingMethod, WebRendererSpec};
+pub mod image {
+    pub use crate::transformations::image_renderer::{ImageSource, ImageSpec, ImageType};
+}
+
+pub mod shader {
+    pub use crate::transformations::shader::ShaderSpec;
+}
+
+pub mod web_renderer {
+    pub use crate::transformations::web_renderer::{
+        WebEmbeddingMethod, WebRendererInitOptions, WebRendererSpec, EMBED_SOURCE_FRAMES_MESSAGE,
+        GET_FRAME_POSITIONS_MESSAGE, UNEMBED_SOURCE_FRAMES_MESSAGE,
+    };
+}
