--- compositor_render/src/sync_renderer.rs
@@ -1,101 +1,0 @@
-use std::sync::{Arc, Mutex};
-
-use crate::{
-    error::{
-        InitRendererEngineError, RegisterRendererError, RenderSceneError, UnregisterRendererError,
-        UpdateSceneError,
-    },
-    event_loop::EventLoop,
-    registry::RegistryType,
-    renderer::{Renderer, RendererOptions},
-    scene::OutputScene,
-    transformations::{
-        image_renderer::{Image, ImageSpec},
-        shader::{Shader, ShaderSpec},
-        web_renderer::{WebRenderer, WebRendererSpec},
-    },
-    FrameSet, InputId, OutputId, RendererId,
-};
-
-/// RendererSpec provides configuration necessary to construct Renderer. Renderers
-/// are entities like shader, image or chromium_instance and can be used by nodes
-/// to transform or generate frames.
-#[derive(Debug)]
-pub enum RendererSpec {
-    Shader(ShaderSpec),
-    WebRenderer(WebRendererSpec),
-    Image(ImageSpec),
-}
-
-#[derive(Clone)]
-pub struct SyncRenderer(Arc<Mutex<Renderer>>);
-
-impl SyncRenderer {
-    pub fn new(opts: RendererOptions) -> Result<(Self, EventLoop), InitRendererEngineError> {
-        let renderer = Renderer::new(opts)?;
-        let event_loop = EventLoop::new(renderer.chromium_context.cef_context());
-
-        Ok((Self(Arc::new(Mutex::new(renderer))), event_loop))
-    }
-
-    pub fn register_renderer(&self, spec: RendererSpec) -> Result<(), RegisterRendererError> {
-        let ctx = self.0.lock().unwrap().register_ctx();
-        let mut guard = self.0.lock().unwrap();
-        match spec {
-            RendererSpec::Shader(spec) => {
-                let shader_id = spec.shader_id.clone();
-
-                match Shader::new(&ctx.wgpu_ctx, spec) {
-                    Ok(shader) => Ok(guard
-                        .renderers
-                        .shaders
-                        .register(shader_id, Arc::new(shader))?),
-                    Err(err) => Err(RegisterRendererError::Shader(err, shader_id)),
-                }
-            }
-            RendererSpec::WebRenderer(params) => {
-                let instance_id = params.instance_id.clone();
-                let web = WebRenderer::new(&ctx, params)
-                    .map_err(|err| RegisterRendererError::Web(err, instance_id.clone()))?;
-
-                Ok(guard
-                    .renderers
-                    .web_renderers
-                    .register(instance_id, Arc::new(web))?)
-            }
-            RendererSpec::Image(spec) => {
-                let image_id = spec.image_id.clone();
-                let asset = Image::new(&ctx, spec)
-                    .map_err(|err| RegisterRendererError::Image(err, image_id.clone()))?;
-
-                Ok(guard.renderers.images.register(image_id, asset)?)
-            }
-        }
-    }
-
-    pub fn unregister_renderer(
-        &self,
-        renderer_id: &RendererId,
-        registry_type: RegistryType,
-    ) -> Result<(), UnregisterRendererError> {
-        let mut guard = self.0.lock().unwrap();
-        // TODO
-        //guard
-        //    .scene_spec
-        //    .validate_can_unregister(renderer_id, registry_type)?;
-        match registry_type {
-            RegistryType::Shader => guard.renderers.shaders.unregister(renderer_id)?,
-            RegistryType::WebRenderer => guard.renderers.web_renderers.unregister(renderer_id)?,
-            RegistryType::Image => guard.renderers.images.unregister(renderer_id)?,
-        }
-        Ok(())
-    }
-
-    pub fn render(&self, input: FrameSet<InputId>) -> Result<FrameSet<OutputId>, RenderSceneError> {
-        self.0.lock().unwrap().render(input)
-    }
-
-    pub fn update_scene(&mut self, scene_specs: Vec<OutputScene>) -> Result<(), UpdateSceneError> {
-        self.0.lock().unwrap().update_scene(scene_specs)
-    }
-}
