+++ compositor_render/src/transformations/web_renderer/chromium_sender.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::{
-    renderer::{render_graph::NodeId, RegisterCtx},
+    state::{render_graph::NodeId, RegisterCtx},
     Resolution,
 };
 use crossbeam_channel::{Receiver, Sender};
