+++ compositor_render/src/transformations/web_renderer/chromium_sender_thread.rs
@@ -10,19 +10,18 @@ use crossbeam_channel::{Receiver, Sender};
 use log::error;
 
 use crate::error::ErrorStack;
-use crate::renderer::render_graph::NodeId;
-use crate::renderer::RegisterCtx;
+use crate::state::render_graph::NodeId;
+use crate::state::RegisterCtx;
 use crate::transformations::web_renderer::chromium_sender::{
     ChromiumSenderMessage, UpdateSharedMemoryInfo,
 };
 use crate::transformations::web_renderer::shared_memory::{SharedMemory, SharedMemoryError};
 use crate::transformations::web_renderer::WebRenderer;
+use crate::wgpu::texture::utils::pad_to_256;
 use crate::Resolution;
-use crate::{
-    wgpu::texture::utils::pad_to_256, EMBED_SOURCE_FRAMES_MESSAGE, GET_FRAME_POSITIONS_MESSAGE,
-};
 
 use super::{browser_client::BrowserClient, chromium_context::ChromiumContext};
+use super::{EMBED_SOURCE_FRAMES_MESSAGE, GET_FRAME_POSITIONS_MESSAGE};
 
 pub(super) struct ChromiumSenderThread {
     chromium_ctx: Arc<ChromiumContext>,
